backend/requirements.txt
itsdangerous==0.24
flask_restx==0.2.0
pandas==1.1.5
Flask==1.0.2
textdistance==4.2.1
scipy==1.5.4
Flask_Cors==3.0.6
python_dateutil==2.8.1
scikit_learn==0.24.1
python-dotenv==0.15.0

